move mode switching buttons to own component
