+++ src/components/author/revision/checkout-reject-buttons.tsx
