+++ src/components/author/revision/revision.tsx
@@ -9,11 +9,12 @@ import dynamic from 'next/dynamic'
 import { tint } from 'polished'
 import * as React from 'react'
 import ReactDiffViewer from 'react-diff-viewer'
-import styled, { css } from 'styled-components'
+import styled from 'styled-components'
 
-import { PageTitle } from '../content/page-title'
-import { UserTools } from '../user-tools/user-tools'
+import { PageTitle } from '../../content/page-title'
+import { UserTools } from '../../user-tools/user-tools'
 import type { CheckoutRejectButtonsProps } from './checkout-reject-buttons'
+import { RevisionModeSwitcher } from './revision-mode-switcher'
 import { useAuthentication } from '@/auth/use-authentication'
 import { useCanDo } from '@/auth/use-can-do'
 import { Geogebra } from '@/components/content/geogebra'
@@ -23,12 +24,7 @@ import { TimeAgo } from '@/components/time-ago'
 import { UserLink } from '@/components/user/user-link'
 import { useInstanceData } from '@/contexts/instance-context'
 import { RevisionData } from '@/data-types'
-import {
-  makePadding,
-  makeTransparentButton,
-  inputFontReset,
-  makeLightButton,
-} from '@/helper/css'
+import { makePadding, inputFontReset, makeLightButton } from '@/helper/css'
 import { getIconByTypename } from '@/helper/icon-by-entity-type'
 import { renderArticle } from '@/schema/article-renderer'
 
@@ -37,13 +33,13 @@ export interface RevisionProps {
 }
 
 const CheckoutRejectButtons = dynamic<CheckoutRejectButtonsProps>(() =>
-  import('@/components/author/checkout-reject-buttons').then(
+  import('@/components/author/revision/checkout-reject-buttons').then(
     (mod) => mod.CheckoutRejectButtons
   )
 )
 
 //current is the checked out revision
-type DisplayMode = 'this' | 'current' | 'compare'
+export type DisplayMode = 'this' | 'current' | 'compare'
 
 export function Revision({ data }: RevisionProps) {
   const auth = useAuthentication()
@@ -65,65 +61,79 @@ export function Revision({ data }: RevisionProps) {
 
   return (
     <>
-      <BackButton href={repositoryAlias} className="mt-6 mx-side">
-        <FontAwesomeIcon icon={faArrowCircleLeft} />{' '}
-        {strings.revisions.toContent}
-      </BackButton>
+      {renderHeader()}
 
-      {renderNotice()}
-      <PageTitle
-        title={
-          data.currentRevision.metaTitle ||
-          data.currentRevision.title ||
-          strings.entities.revision
-        }
-        headTitle
-        icon={icon ? icon : undefined}
-      />
-      {renderUserTools(true)}
-      <p className="serlo-p leading-7">
-        {data.changes && (
-          <>
-            <b>{strings.revisions.changes}:</b> {data.changes}
-          </>
+      <main>
+        {dataSet.title !== undefined && (
+          <PreviewBox title={strings.revisions.title} diffType="title">
+            <h1 className="serlo-h1">{dataSet.title}</h1>
+          </PreviewBox>
         )}
-        <br />
-        {strings.revisions.by} <UserLink user={data.user} />{' '}
-        <TimeAgo datetime={new Date(data.date)} dateAsTitle />
-      </p>
-
-      <MetaBar>{renderButtons()}</MetaBar>
-
-      {dataSet.title !== undefined && (
-        <PreviewBox title={strings.revisions.title} diffType="title">
-          <h1 className="serlo-h1">{dataSet.title}</h1>
-        </PreviewBox>
-      )}
-      {dataSet.content !== undefined && (
-        <PreviewBox title={strings.revisions.content} diffType="content">
-          {renderArticle(
-            dataSet.content || [],
-            `revision${dataSet.id || 'empty'}`
-          )}
-        </PreviewBox>
-      )}
-      {renderVideoOrAppletBox()}
-      {dataSet.metaTitle && (
-        <PreviewBox title={strings.revisions.metaTitle} diffType="metaTitle">
-          {dataSet.metaTitle}
-        </PreviewBox>
-      )}
-      {dataSet.metaDescription && (
-        <PreviewBox
-          title={strings.revisions.metaDescription}
-          diffType="metaDescription"
-        >
-          {dataSet.metaDescription}
-        </PreviewBox>
-      )}
+        {dataSet.content !== undefined && (
+          <PreviewBox title={strings.revisions.content} diffType="content">
+            {renderArticle(
+              dataSet.content || [],
+              `revision${dataSet.id || 'empty'}`
+            )}
+          </PreviewBox>
+        )}
+        {renderVideoOrAppletBox()}
+        {dataSet.metaTitle && (
+          <PreviewBox title={strings.revisions.metaTitle} diffType="metaTitle">
+            {dataSet.metaTitle}
+          </PreviewBox>
+        )}
+        {dataSet.metaDescription && (
+          <PreviewBox
+            title={strings.revisions.metaDescription}
+            diffType="metaDescription"
+          >
+            {dataSet.metaDescription}
+          </PreviewBox>
+        )}
+      </main>
       {renderUserTools(false)}
     </>
   )
+
+  function renderHeader() {
+    return (
+      <header>
+        <BackButton href={repositoryAlias} className="mt-6 mx-side">
+          <FontAwesomeIcon icon={faArrowCircleLeft} />{' '}
+          {strings.revisions.toContent}
+        </BackButton>
+
+        {renderNotice()}
+        <PageTitle
+          title={
+            data.currentRevision.metaTitle ||
+            data.currentRevision.title ||
+            strings.entities.revision
+          }
+          headTitle
+          icon={icon ? icon : undefined}
+        />
+        {renderUserTools(true)}
+        <p className="serlo-p leading-7">
+          {data.changes && (
+            <>
+              <b>{strings.revisions.changes}:</b> {data.changes}
+            </>
+          )}
+          <br />
+          {strings.revisions.by} <UserLink user={data.user} />{' '}
+          <TimeAgo datetime={new Date(data.date)} dateAsTitle />
+        </p>
+
+        <RevisionModeSwitcher
+          isCurrent={isCurrentRevision}
+          setDisplayMode={setDisplayMode}
+          displayMode={displayMode}
+        />
+      </header>
+    )
+  }
 
   function renderUserTools(above: boolean) {
     return (
@@ -145,31 +155,6 @@ export function Revision({ data }: RevisionProps) {
             ) : undefined,
         }}
       />
-    )
-  }
-
-  function renderButtons() {
-    return (
-      <>
-        {!isCurrentRevision &&
-          renderButton('compare', strings.revisions.compare)}
-        {renderButton('current', strings.revisions.currentVersion)}
-        {!isCurrentRevision &&
-          renderButton('this', strings.revisions.thisVersion)}
-      </>
-    )
-  }
-
-  function renderButton(mode: DisplayMode, title: string) {
-    //blur-hack, use https://caniuse.com/#feat=css-focus-visible when supported
-    return (
-      <Button
-        onPointerUp={(e) => e.currentTarget.blur()}
-        onClick={() => setDisplayMode(mode)}
-        current={displayMode === mode}
-      >
-        {title}
-      </Button>
     )
   }
 
@@ -280,21 +265,6 @@ const Box = styled.div<{ withPadding?: boolean }>`
   }
 `
 
-const Button = styled.button<{ current?: boolean }>`
-  ${makeTransparentButton};
-  margin-left: 5px;
-
-  ${(props) =>
-    props.current &&
-    css`
-      &,
-      &:hover {
-        background-color: ${(props) => props.theme.colors.brand};
-        color: #fff;
-      }
-    `}
-`
-
 const BackButton = styled(Link)`
   ${makeLightButton}
 
@@ -303,19 +273,6 @@ const BackButton = styled(Link)`
     margin-right: 2px;
     padding-top: 1px;
   }
-`
-
-const MetaBar = styled.div`
-  ${makePadding};
-  display: flex;
-  justify-content: flex-end;
-  position: sticky;
-  z-index: 50;
-  padding-top: 25px;
-  padding-bottom: 10px;
-  margin-bottom: 20px;
-  top: 0;
-  background-color: #fff;
 `
 
 const DiffViewerWrapper = styled.div`
